clear history has landed
