+++ src/client/components/containers/HistoryContainer.jsx
@@ -2,7 +2,9 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import * as actions from '../../actions/actions';
 import HistoryDate from '../display/HistoryDate.jsx';
+import ClearHistoryBtn from '../display/ClearHistoryBtn.jsx';
 import historyController from '../../controllers/historyController';
+
 const { dialog } = require('electron').remote;
 import parse from 'date-fns/parse'
 
@@ -46,20 +48,7 @@ class HistoryContainer extends Component {
       <div className={'historyDate-container'}>
         <h1>History</h1>
         <div className="clear-history">
-          <button onClick={() => {
-            const opts = {
-              type: 'warning', 
-              buttons: ['Okay', 'Cancel'],
-              message: 'Are you sure you want to clear history?'
-            }
-
-            dialog.showMessageBox(null, opts, (response) => {
-              if (response === 0) {
-                this.props.clearHistory();
-                historyController.clearHistoryFromIndexedDb();
-              }
-            })
-          }}>Clear History</button>
+          <ClearHistoryBtn clearHistory={this.props.clearHistory} />
         </div>
         {historyDates}
       </div>
